feat(store): include magnet total size in response

+++ internal/magnet_cache/db.go
@@ -22,7 +22,7 @@ var mcLog = logger.Scoped(TableName)
 type File struct {
 	Idx  int    `json:"i"`
 	Name string `json:"n"`
-	Size int    `json:"s"`
+	Size int64  `json:"s"`
 	SId  string `json:"-"`
 }
